front: synchro split between manchette and get
Main modifications:
- Updates useManchetteWithSpaceTimeChart and styles, so that all
  elements in manchette are positioned absolutely using proper helpers
  from the SpaceTimeChart
- Renames computeWaypointsToDisplay as selectWaypointsToDisplay as it
  now only selects the waypoints to display, and no more handles their
  heights (and updates unit tests accordingly)
- Adds optional splitPoints argument to useManchetteWithSpaceTimeChart
- Adds new story about rendering split sections using
  useManchetteWithSpaceTimeChart

Collateral modifications:
- Cleans and types useManchetteWithSpaceTimeChart outputs
  (manchetteProps are only Manchette props, spaceTimeChartProps are
  only SpaceTimeChart props, etc...), and updates all existing stories
  accordingly
- Fixes various alignment issues between SpaceTimeChart and manchette
  (2 pixels trimmed from SpaceTimeChart on top, magic numbers to handle
  bottom alignment instead of correcting the GET CAPTION_SIZE)
- Removes WaypointList from Manchette, and instead directly inserts
  waypoints and split sections from the Manchette component
