+++ storybook/stories/ui-charts/manchetteWithSpaceTimeChart/rectangle-zoom.stories.tsx
@@ -1,13 +1,14 @@
 import React, { useRef } from 'react';
 
 import {
-  PathLayer,
   SpaceTimeChart,
   Manchette,
   useManchetteWithSpaceTimeChart,
   type ProjectPathTrainResult,
   type Waypoint,
   ZoomRect,
+  PathLayer,
+  usePaths,
 } from '@osrd-project/ui-charts';
 import { Slider } from '@osrd-project/ui-core';
 import { ZoomIn } from '@osrd-project/ui-icons';
@@ -36,9 +37,11 @@ const ManchetteWithSpaceTimeWrapper = ({
 }: ManchetteWithSpaceTimeWrapperProps) => {
   const manchetteWithSpaceTimeChartRef = useRef<HTMLDivElement>(null);
   const spaceTimeChartRef = useRef<HTMLDivElement>(null);
+  const paths = usePaths(projectPathTrainResult);
   const {
     manchetteProps,
     spaceTimeChartProps,
+    rect,
     handleScroll,
     toggleZoomMode,
     zoomMode,
@@ -48,13 +51,12 @@ const ManchetteWithSpaceTimeWrapper = ({
     spaceScale,
   } = useManchetteWithSpaceTimeChart({
     waypoints,
-    projectPathTrainResult,
     manchetteWithSpaceTimeChartRef,
     height: DEFAULT_HEIGHT,
     spaceTimeChartRef,
     defaultTimeOrigin: Math.min(...projectPathTrainResult.map((p) => +p.departureTime)),
   });
-  const selectedPath = spaceTimeChartProps.paths[selectedTrain].id;
+  const selectedPath = paths[selectedTrain].id;
   return (
     <div className="manchette-space-time-chart-wrapper">
       <div
@@ -68,11 +70,7 @@ const ManchetteWithSpaceTimeWrapper = ({
         onScroll={handleScroll}
       >
         <Manchette {...manchetteProps} />
-        <div
-          className="space-time-chart-container w-full sticky"
-          ref={spaceTimeChartRef}
-          style={{ bottom: 0, left: 0, top: 2, height: `${DEFAULT_HEIGHT - 6}px` }}
-        >
+        <div className="space-time-chart-container w-full sticky" ref={spaceTimeChartRef}>
           <div className="toolbar">
             <button
               type="button"
@@ -88,7 +86,7 @@ const ManchetteWithSpaceTimeWrapper = ({
             })}
             {...spaceTimeChartProps}
           >
-            {spaceTimeChartProps.paths.map((path) => (
+            {paths.map((path) => (
               <PathLayer
                 key={path.id}
                 path={path}
@@ -96,7 +94,7 @@ const ManchetteWithSpaceTimeWrapper = ({
                 level={path.id === selectedPath ? 1 : 2}
               />
             ))}
-            {spaceTimeChartProps.rect && <ZoomRect {...spaceTimeChartProps.rect} />}
+            {rect && <ZoomRect {...rect} />}
             <MouseTracker />
           </SpaceTimeChart>
         </div>
