+++ ui-charts/src/manchette/hooks/useManchetteWithSpaceTimeChart.tsx
@@ -1,0 +1,677 @@
+import React, { type ReactNode, useCallback, useEffect, useMemo, useState } from 'react';
+
+import { sortBy, clamp } from 'lodash';
+
+import { type SpaceScale, type SpaceTimeChartProps } from '../../spaceTimeChart';
+import { getSpaceToPixel, spaceScalesToBinaryTree } from '../../spaceTimeChart/utils/scales';
+import type { ManchetteProps } from '../components/Manchette';
+import {
+  MAX_ZOOM_Y,
+  MIN_ZOOM_Y,
+  ZOOM_Y_DELTA,
+  DEFAULT_ZOOM_MS_PER_PX,
+  MAX_ZOOM_MS_PER_PX,
+  MIN_ZOOM_MS_PER_PX,
+  BASE_WAYPOINT_HEIGHT,
+  FOOTER_HEIGHT,
+  WAYPOINT_LINE_HEIGHT,
+} from '../consts';
+import type { InteractiveWaypoint, Waypoint } from '../types';
+import { getDistance, calcTotalDistance } from '../utils';
+import {
+  selectWaypointsToDisplay,
+  getScales,
+  timeScaleToZoomValue,
+  spaceScaleToZoomValue,
+  getExtremaScales,
+  zoomValueToSpaceScale,
+  zoomValueToTimeScale,
+  zoomX,
+} from '../utils/helpers';
+
+const WAYPOINTS_OFFSET = 16;
+
+type State = {
+  xZoom: number;
+  yZoom: number;
+  timeOrigin: number;
+  spaceOrigin: number;
+  /** current x PIXEL offset from x origin */
+  xOffset: number;
+  /** current y PIXEL offset from y origin (the current y-scroll of the view. always updates) */
+  yOffset: number;
+  /** only update after a zoom. used to update back the view scroll value */
+  scrollTo: number | null;
+  panning: { initialOffset: { x: number; y: number } } | null;
+  zoomMode: boolean;
+  rect: {
+    timeStart: Date;
+    timeEnd: Date;
+    spaceStart: number; // mm
+    spaceEnd: number; // mm
+  } | null;
+  pixelRect: {
+    xStart: number;
+    xEnd: number;
+    yStart: number;
+    yEnd: number;
+  } | null;
+  isProportional: boolean;
+  waypointsChart: Waypoint[];
+  scales: SpaceScale[];
+};
+
+export type SplitPoint = {
+  /** helper identify nodes for this split point in the React tree */
+  id: string;
+  /** position of this split point, in mm from start */
+  position: number;
+  /** size of this split point, in pixels on screen */
+  size: number;
+  /** a React node to render the split point, space/time chart side */
+  spaceTimeChartNode?: ReactNode;
+  /** a React node to render the split point, manchette side */
+  manchetteNode?: ReactNode;
+};
+
+const useManchetteWithSpaceTimeChart = ({
+  waypoints,
+  manchetteWithSpaceTimeChartRef,
+  height = 561,
+  spaceTimeChartRef,
+  defaultTimeOrigin = 0,
+  defaultSpaceOrigin = 0,
+  splitPoints = [],
+}: {
+  waypoints: Waypoint[];
+  manchetteWithSpaceTimeChartRef: React.RefObject<HTMLDivElement>;
+  height?: number;
+  spaceTimeChartRef?: React.RefObject<HTMLDivElement>;
+  defaultTimeOrigin?: number;
+  defaultSpaceOrigin?: number;
+  splitPoints?: SplitPoint[];
+}) => {
+  const [isShiftPressed, setIsShiftPressed] = useState(false);
+  const [state, setState] = useState<State>({
+    xZoom: timeScaleToZoomValue(DEFAULT_ZOOM_MS_PER_PX),
+    yZoom: 1,
+    timeOrigin: defaultTimeOrigin,
+    spaceOrigin: defaultSpaceOrigin,
+    xOffset: 0,
+    yOffset: 0,
+    scrollTo: null,
+    panning: null,
+    zoomMode: false,
+    rect: null,
+    pixelRect: null,
+    isProportional: true,
+    waypointsChart: [],
+    scales: [],
+  });
+
+  const {
+    xZoom,
+    yZoom,
+    timeOrigin,
+    spaceOrigin,
+    xOffset,
+    yOffset,
+    scrollTo,
+    panning,
+    zoomMode,
+    rect,
+    pixelRect,
+    isProportional,
+  } = state;
+
+  /** used when we change the train dataset for example to center the chart on the new data */
+  const setTimeOrigin = useCallback((newTimeOrigin: number) => {
+    setState((prev) => ({ ...prev, timeOrigin: newTimeOrigin }));
+  }, []);
+
+  const canvasDrawingHeight = height - FOOTER_HEIGHT; // 521
+  const drawingHeightWithoutTopPadding = canvasDrawingHeight - BASE_WAYPOINT_HEIGHT / 2; // 505
+  const drawingHeightWithoutBothPadding = canvasDrawingHeight - BASE_WAYPOINT_HEIGHT; // 489
+  const totalDistance = calcTotalDistance(waypoints);
+
+  const { minZoomMillimeterPerPx, maxZoomMillimeterPerPx } = getExtremaScales(
+    drawingHeightWithoutTopPadding,
+    drawingHeightWithoutBothPadding,
+    totalDistance
+  );
+
+  const handleRectangleZoom = useCallback(
+    ({
+      scales: { chosenTimeScale, chosenSpaceScale },
+      overrideState,
+    }: {
+      scales: { chosenTimeScale: number; chosenSpaceScale?: number };
+      overrideState?: Partial<State>;
+    }) => {
+      setState((prev) => {
+        if (prev.zoomMode || !prev.rect) {
+          return prev;
+        }
+        const newTimeScale = clamp(chosenTimeScale, MAX_ZOOM_MS_PER_PX, MIN_ZOOM_MS_PER_PX);
+        const timeZoomValue = timeScaleToZoomValue(newTimeScale);
+        const leftRectSide = Math.min(Number(prev.rect.timeStart), Number(prev.rect.timeEnd));
+        const newXOffset = (timeOrigin - leftRectSide) / newTimeScale;
+
+        let newYZoom = yZoom;
+        let newYOffset = yOffset;
+        if (chosenSpaceScale) {
+          const newSpaceScale = clamp(
+            chosenSpaceScale,
+            maxZoomMillimeterPerPx,
+            minZoomMillimeterPerPx
+          );
+          // we don’t need to handle this case and compute an offset
+          // this condition happens when we draw a rectangle
+          // larger than the entire chart (minus padding)
+          // that would actually zoom OUT if it wasn’t clamped.
+          if (newSpaceScale !== minZoomMillimeterPerPx) {
+            newYZoom = spaceScaleToZoomValue(
+              minZoomMillimeterPerPx,
+              maxZoomMillimeterPerPx,
+              newSpaceScale
+            );
+            const topRectSide = Math.min(prev.rect.spaceStart, prev.rect.spaceEnd);
+            newYOffset = Math.abs(spaceOrigin - topRectSide) / newSpaceScale;
+          }
+        }
+
+        return {
+          ...prev,
+          xZoom: timeZoomValue,
+          yZoom: newYZoom,
+          xOffset: newXOffset,
+          yOffset: newYOffset,
+          scrollTo: newYOffset,
+          ...overrideState,
+        };
+      });
+    },
+    [timeOrigin, spaceOrigin, minZoomMillimeterPerPx, maxZoomMillimeterPerPx, yOffset, yZoom]
+  );
+
+  useEffect(() => {
+    if (rect && !zoomMode && spaceTimeChartRef?.current) {
+      const { timeStart, timeEnd, spaceStart, spaceEnd } = rect;
+      const timeRange = Math.abs(Number(timeEnd) - Number(timeStart)); // width of rect in ms
+      const spaceRange = Math.abs(spaceEnd - spaceStart); // height of rect in mm
+
+      const chosenTimeScale = timeRange / spaceTimeChartRef.current.clientWidth;
+      if (isProportional) {
+        const chosenSpaceScale = spaceRange / drawingHeightWithoutTopPadding;
+        handleRectangleZoom({
+          scales: { chosenTimeScale, chosenSpaceScale },
+          overrideState: { rect: null },
+        });
+      } else if (pixelRect) {
+        const currentStopHeight = BASE_WAYPOINT_HEIGHT * yZoom;
+        const { yStart, yEnd } = pixelRect;
+        const numberOfStopsInRect = Math.abs(yEnd - yStart) / currentStopHeight;
+        let newStopHeight = drawingHeightWithoutTopPadding / numberOfStopsInRect;
+        // at maximum zoom, we want 3 stops displayed
+        const maxStopHeight =
+          drawingHeightWithoutTopPadding / (2 + BASE_WAYPOINT_HEIGHT / newStopHeight);
+        newStopHeight = Math.min(newStopHeight, maxStopHeight);
+        const newYZoom = newStopHeight / BASE_WAYPOINT_HEIGHT;
+        const rectTop = yOffset + Math.min(yStart, yEnd) - WAYPOINT_LINE_HEIGHT;
+        const numberOfStopsBeforeRectTop = rectTop / currentStopHeight;
+        const newYOffset = numberOfStopsBeforeRectTop * newStopHeight;
+
+        handleRectangleZoom({
+          scales: {
+            chosenTimeScale,
+          },
+          overrideState: {
+            rect: null,
+            pixelRect: null,
+            yZoom: newYZoom,
+            yOffset: newYOffset,
+            scrollTo: newYOffset,
+          },
+        });
+      }
+    }
+  }, [
+    state.rect,
+    state.zoomMode,
+    handleRectangleZoom,
+    drawingHeightWithoutTopPadding,
+    isProportional,
+    pixelRect,
+    rect,
+    spaceTimeChartRef,
+    yOffset,
+    yZoom,
+    zoomMode,
+  ]);
+
+  const zoomYIn = useCallback(() => {
+    const maxZoom = isProportional
+      ? MAX_ZOOM_Y
+      : (drawingHeightWithoutTopPadding - BASE_WAYPOINT_HEIGHT) / (2 * BASE_WAYPOINT_HEIGHT);
+    const newYZoom = Math.min(yZoom + ZOOM_Y_DELTA, maxZoom);
+    if (newYZoom !== yZoom) {
+      const newYOffset = yOffset * (newYZoom / yZoom);
+
+      setState((prev) => ({
+        ...prev,
+        yZoom: newYZoom,
+        yOffset: newYOffset,
+        scrollTo: newYOffset,
+      }));
+    }
+  }, [yZoom, yOffset, drawingHeightWithoutTopPadding, isProportional]);
+
+  const zoomYOut = useCallback(() => {
+    const newYZoom = Math.max(MIN_ZOOM_Y, yZoom - ZOOM_Y_DELTA);
+    if (newYZoom !== yZoom) {
+      const newYOffset = yOffset * (newYZoom / yZoom);
+      setState((prev) => ({
+        ...prev,
+        yZoom: newYZoom,
+        yOffset: newYOffset,
+        scrollTo: newYOffset,
+      }));
+    }
+  }, [yZoom, yOffset]);
+
+  useEffect(() => {
+    if (scrollTo !== null && manchetteWithSpaceTimeChartRef.current) {
+      manchetteWithSpaceTimeChartRef.current.scrollTo({
+        top: scrollTo,
+        behavior: 'instant',
+      });
+    }
+  }, [scrollTo, manchetteWithSpaceTimeChartRef]);
+
+  const resetZoom = useCallback(() => {
+    setState((prev) => ({ ...prev, yZoom: 1 }));
+  }, []);
+
+  const handleScroll = useCallback(() => {
+    if (!isShiftPressed && manchetteWithSpaceTimeChartRef.current) {
+      const { scrollTop } = manchetteWithSpaceTimeChartRef.current;
+      if (scrollTop || scrollTop === 0) {
+        setState((prev) => ({ ...prev, yOffset: scrollTop }));
+      }
+    }
+  }, [isShiftPressed, manchetteWithSpaceTimeChartRef]);
+
+  const handleKeyDown = useCallback((event: KeyboardEvent) => {
+    if (event.key === 'Shift') {
+      setIsShiftPressed(true);
+    }
+  }, []);
+
+  const handleKeyUp = useCallback((event: KeyboardEvent) => {
+    if (event.key === 'Shift') {
+      setIsShiftPressed(false);
+    }
+  }, []);
+
+  useEffect(() => {
+    window.addEventListener('keydown', handleKeyDown);
+    window.addEventListener('keyup', handleKeyUp);
+
+    return () => {
+      window.removeEventListener('keydown', handleKeyDown);
+      window.removeEventListener('keyup', handleKeyUp);
+    };
+  }, [handleKeyDown, handleKeyUp]);
+
+  const toggleMode = useCallback(() => {
+    setState((prev) => ({ ...prev, isProportional: !prev.isProportional }));
+  }, []);
+
+  const toggleZoomMode = useCallback(() => {
+    setState((prev) => ({ ...prev, zoomMode: !prev.zoomMode }));
+  }, []);
+
+  const handleXZoom = useCallback(
+    (newXZoom: number, xPosition = (spaceTimeChartRef?.current?.offsetWidth || 0) / 2) => {
+      setState((prev) => ({
+        ...prev,
+        ...zoomX(prev.xZoom, prev.xOffset, newXZoom, xPosition),
+      }));
+    },
+    [spaceTimeChartRef]
+  );
+
+  const spaceScales = useMemo(() => {
+    // Here, we first compute the base scales, and then we insert a flat step for each split point:
+    const baseScales = getScales(
+      waypoints,
+      { height, isProportional, yZoom },
+      minZoomMillimeterPerPx,
+      maxZoomMillimeterPerPx
+    );
+
+    if (!splitPoints) return baseScales;
+
+    // Constant scale:
+    if (isProportional) {
+      const baseScale = baseScales[0];
+      const coefficient: number =
+        'coefficient' in baseScale && typeof baseScale.coefficient === 'number'
+          ? baseScale.coefficient
+          : (baseScale.to - baseScale.from) / baseScale.size;
+
+      return splitPoints
+        .flatMap(({ position, size: splitPointHeight }) => [
+          {
+            to: position,
+            coefficient,
+          },
+          {
+            to: position,
+            size: splitPointHeight,
+          },
+        ])
+        .concat({
+          to: baseScales.at(-1)!.to,
+          coefficient,
+        });
+    }
+
+    // Varying scales:
+    const allScales: SpaceScale[] = [];
+    baseScales.forEach((baseScale) => {
+      const coefficient =
+        'coefficient' in baseScale && typeof baseScale.coefficient === 'number'
+          ? baseScale.coefficient
+          : (baseScale.to - baseScale.from) / baseScale.size;
+
+      // Search for split points strictly BETWEEN from and to:
+      const relevantPoints = splitPoints.filter(
+        (point) => point.position > baseScale.from && point.position < baseScale.to
+      );
+
+      relevantPoints.forEach(({ position }) => {
+        allScales.push({
+          to: position,
+          coefficient,
+        });
+      });
+
+      allScales.push({
+        to: baseScale.to,
+        coefficient,
+      });
+    });
+
+    splitPoints.forEach((point) => {
+      allScales.push({
+        to: point.position,
+        size: point.size,
+      });
+    });
+
+    return sortBy(allScales, (scale) => scale.to + ('size' in scale ? Number.EPSILON : 0));
+  }, [
+    waypoints,
+    height,
+    isProportional,
+    yZoom,
+    minZoomMillimeterPerPx,
+    maxZoomMillimeterPerPx,
+    splitPoints,
+  ]);
+
+  const waypointsWithoutSplitPoints = useMemo(() => {
+    const splitPointPositions = new Set(splitPoints?.map((point) => point.position) || []);
+    const filteredWaypoints = selectWaypointsToDisplay(waypoints, {
+      height,
+      isProportional,
+      yZoom,
+    });
+    return filteredWaypoints.filter((waypoint) => !splitPointPositions.has(waypoint.position));
+  }, [splitPoints, waypoints, height, isProportional, yZoom]);
+
+  const { manchetteContents, manchetteHeight } = useMemo(() => {
+    const spaceScaleTree = spaceScalesToBinaryTree(spaceOrigin, spaceScales);
+    const getSpacePixel = getSpaceToPixel(0, spaceScaleTree);
+    const totalManchetteHeight = getSpacePixel(spaceScales.at(-1)!.to, true) + BASE_WAYPOINT_HEIGHT;
+
+    if (!splitPoints)
+      return {
+        manchetteHeight: totalManchetteHeight,
+        manchetteContents: waypointsWithoutSplitPoints,
+      };
+
+    // Identify all manchette contents (split sections and waypoints):
+    let allContents: (
+      | { type: 'waypoint'; position: number; waypoint: Waypoint }
+      | { type: 'splitSection'; position: number; split: SplitPoint }
+    )[] = waypointsWithoutSplitPoints.map((wp) => ({
+      type: 'waypoint',
+      waypoint: wp,
+      position: wp.position,
+    }));
+    allContents = allContents.concat(
+      splitPoints.map((sp) => ({
+        type: 'splitSection',
+        split: sp,
+        position: sp.position,
+      }))
+    );
+
+    // Sort all contents by position:
+    const allSortedContents = sortBy(allContents, 'position');
+
+    // Iterate over all contents, to set each split section's style, and correct waypoints styles
+    // accordingly:
+    const finalContents: (InteractiveWaypoint | ReactNode)[] = [];
+    allSortedContents.forEach((content) => {
+      switch (content.type) {
+        case 'waypoint': {
+          finalContents.push({
+            ...content.waypoint,
+            styles: {
+              position: 'absolute',
+              top: `${getSpacePixel(content.position)}px`,
+              height: `${BASE_WAYPOINT_HEIGHT}px`,
+            },
+          });
+          break;
+        }
+        case 'splitSection': {
+          const {
+            split: { position },
+          } = content;
+
+          const top = getSpacePixel(position);
+          finalContents.push(
+            <div
+              style={{
+                // This WAYPOINTS_OFFSET correction is there because:
+                // - Waypoints are positioned so that their lines (at WAYPOINTS_OFFSET pixels from
+                //   the top) are aligned with SpaceTimeChart's space lines
+                // - Split points are positioned so that their top is aligned with SpaceTimeChart's
+                //   space lines
+                top: `${top + WAYPOINTS_OFFSET}px`,
+                height: `${getSpacePixel(position, true) - top}px`,
+                position: 'absolute',
+                width: '100%',
+              }}
+            >
+              {content.split.manchetteNode}
+            </div>
+          );
+          break;
+        }
+      }
+    });
+
+    return {
+      manchetteHeight: totalManchetteHeight,
+      manchetteContents: finalContents,
+    };
+  }, [spaceOrigin, spaceScales, splitPoints, waypointsWithoutSplitPoints]);
+
+  return useMemo<{
+    manchetteProps: ManchetteProps;
+    spaceTimeChartProps: SpaceTimeChartProps;
+    handleScroll: () => void;
+    handleXZoom: (newXZoom: number, xPosition?: number) => void;
+    xZoom: number;
+    toggleZoomMode: () => void;
+    zoomMode: boolean;
+    rect: State['rect'];
+    timeScale: number;
+    spaceScale: number;
+    setTimeOrigin: (v: number) => void;
+  }>(
+    () => ({
+      manchetteProps: {
+        contents: manchetteContents,
+        height: manchetteHeight,
+        zoomYIn,
+        zoomYOut,
+        resetZoom,
+        toggleMode,
+        yZoom,
+        isProportional,
+        yOffset,
+      },
+      spaceTimeChartProps: {
+        operationalPoints: waypointsWithoutSplitPoints.map((waypoint) => ({
+          ...waypoint,
+          importanceLevel: 1,
+        })),
+        additionalChildren: splitPoints.map((sp) => sp.spaceTimeChartNode),
+        timeScale: zoomValueToTimeScale(xZoom),
+        xOffset,
+        yOffset: -yOffset + WAYPOINTS_OFFSET,
+        timeOrigin,
+        spaceOrigin,
+        spaceScales,
+        onZoom: ({
+          delta,
+          position,
+        }: Parameters<NonNullable<SpaceTimeChartProps['onZoom']>>[0]) => {
+          if (isShiftPressed) {
+            handleXZoom(xZoom + delta, position.x);
+          }
+        },
+        onPan: (payload: Parameters<NonNullable<SpaceTimeChartProps['onPan']>>[0]) => {
+          const {
+            initialData,
+            data,
+            initialPosition,
+            position,
+            isPanning,
+            context: { width, getData },
+          } = payload;
+          const diff = getDistance(initialPosition, position);
+          setState((prev) => {
+            if (!isPanning) {
+              return {
+                ...prev,
+                panning: null,
+                zoomMode: false,
+              };
+            }
+
+            if (zoomMode) {
+              const minPoint = getData({ x: 0, y: 0 });
+              const maxPoint = getData({ x: width, y: canvasDrawingHeight });
+              const timeStart = clamp(initialData.time, minPoint.time, maxPoint.time);
+              const timeEnd = clamp(data.time, minPoint.time, maxPoint.time);
+              const spaceStart = clamp(initialData.position, minPoint.position, maxPoint.position);
+              const spaceEnd = clamp(data.position, minPoint.position, maxPoint.position);
+              const newRect: State['rect'] = {
+                timeStart: new Date(timeStart),
+                timeEnd: new Date(timeEnd),
+                spaceStart,
+                spaceEnd,
+              };
+
+              let newPixelRect: State['pixelRect'] = null;
+              if (!isProportional) {
+                const xStart = clamp(initialPosition.x, 0, width);
+                const xEnd = clamp(position.x, 0, width);
+                const yStart = clamp(initialPosition.y, 0, canvasDrawingHeight);
+                const yEnd = clamp(position.y, 0, canvasDrawingHeight);
+                newPixelRect = { xStart, xEnd, yStart, yEnd };
+              }
+
+              return {
+                ...prev,
+                rect: newRect,
+                pixelRect: newPixelRect,
+              };
+            }
+
+            if (!panning) {
+              return {
+                ...prev,
+                panning: { initialOffset: { x: xOffset, y: yOffset } },
+              };
+            }
+
+            const newState = { ...prev };
+            const { initialOffset } = panning;
+            newState.xOffset = initialOffset.x + diff.x;
+
+            const newYPos = initialOffset.y - diff.y;
+            if (
+              manchetteWithSpaceTimeChartRef.current &&
+              newYPos >= 0 &&
+              newYPos + manchetteWithSpaceTimeChartRef.current.offsetHeight <
+                manchetteWithSpaceTimeChartRef.current.scrollHeight
+            ) {
+              newState.yOffset = newYPos;
+              manchetteWithSpaceTimeChartRef.current.scrollTop = newYPos;
+            }
+            return newState;
+          });
+        },
+      },
+      handleScroll,
+      handleXZoom,
+      xZoom,
+      toggleZoomMode,
+      zoomMode,
+      rect,
+      timeScale: zoomValueToTimeScale(xZoom),
+      spaceScale: zoomValueToSpaceScale(minZoomMillimeterPerPx, maxZoomMillimeterPerPx, yZoom),
+      setTimeOrigin,
+    }),
+    [
+      manchetteContents,
+      manchetteHeight,
+      zoomYIn,
+      zoomYOut,
+      resetZoom,
+      toggleMode,
+      yZoom,
+      isProportional,
+      yOffset,
+      waypointsWithoutSplitPoints,
+      xZoom,
+      xOffset,
+      timeOrigin,
+      spaceOrigin,
+      spaceScales,
+      splitPoints,
+      handleScroll,
+      handleXZoom,
+      toggleZoomMode,
+      zoomMode,
+      rect,
+      minZoomMillimeterPerPx,
+      maxZoomMillimeterPerPx,
+      setTimeOrigin,
+      isShiftPressed,
+      panning,
+      manchetteWithSpaceTimeChartRef,
+      canvasDrawingHeight,
+    ]
+  );
+};
+
+export default useManchetteWithSpaceTimeChart;
