+++ ui-charts/src/manchette/styles/main.css
@@ -3,5 +3,4 @@
 @import 'tailwindcss/utilities';
 @import 'manchette.css';
 @import 'waypoint.css';
-@import 'waypoint-list.css';
 @import 'menu.css';
