--- ui-charts/src/manchette/styles/waypoint-list.css
@@ -1,15 +1,0 @@
-.waypoint-list {
-  width: 349px;
-  border-bottom-left-radius: 0.25rem;
-  padding-inline: 0.5rem 1.625rem;
-
-  @apply bg-white-100;
-
-  .waypoint-wrapper:last-child {
-    .waypoint {
-      &::after {
-        bottom: 16px;
-      }
-    }
-  }
-}
